backend/requirements.txt
fastapi
uvicorn
pdf2image
pillow
python-multipart
PyMuPDF
